+++ src/main/kotlin/day20/Day20.kt
@@ -5,19 +5,19 @@ import base.Day
 class Day20(dayNumber: Int, loadDemoData: Boolean, postFix: String) : Day(dayNumber, loadDemoData, postFix) {
   /**
    * Berechnung der ersten Teilaufgabe
-   * @returns {Int}
+   * @returns {Long}
    * @override
    */
-  override fun calcPartOne(): Int {
-    return 0
+  override fun calcPartOne(): Long {
+    return 0L
   }
 
   /**
    * Berechnung der zweiten Teilaufgabe
-   * @returns {Int}
+   * @returns {Long}
    * @override
    */
-  override fun calcPartTwo(): Int {
-    return 0
+  override fun calcPartTwo(): Long {
+    return 0L
   }
 }
