+++ src/main/kotlin/day23/Day23.kt
@@ -5,11 +5,11 @@ import base.Day
 class Day23(dayNumber: Int, loadDemoData: Boolean) : Day(dayNumber, loadDemoData) {
   /**
    * Berechnung der ersten Teilaufgabe
-   * @returns {Int}
+   * @returns {Long}
    * @override
    */
-  override fun calcPartOne(): Int {
-    return 0
+  override fun calcPartOne(): Long {
+    return 0L
   }
 
   /**
@@ -17,7 +17,7 @@ class Day23(dayNumber: Int, loadDemoData: Boolean) : Day(dayNumber, loadDemoData
    * @returns {Int}
    * @override
    */
-  override fun calcPartTwo(): Int {
-    return 0
+  override fun calcPartTwo(): Long {
+    return 0L
   }
 }
