+++ src/main/kotlin/day25/Day25.kt
@@ -5,19 +5,19 @@ import base.Day
 class Day25(dayNumber: Int, loadDemoData: Boolean) : Day(dayNumber, loadDemoData) {
   /**
    * Berechnung der ersten Teilaufgabe
-   * @returns {Int}
+   * @returns {Long}
    * @override
    */
-  override fun calcPartOne(): Int {
+  override fun calcPartOne(): Long {
     return 0
   }
 
   /**
    * Berechnung der zweiten Teilaufgabe
-   * @returns {Int}
+   * @returns {Long}
    * @override
    */
-  override fun calcPartTwo(): Int {
+  override fun calcPartTwo(): Long {
     return 0
   }
 }
