+++ src/main/kotlin/template/DayXX.kt
@@ -5,19 +5,19 @@ import base.Day
 class DayXX(dayNumber: Int, loadDemoData: Boolean) : Day(dayNumber, loadDemoData) {
   /**
    * Berechnung der ersten Teilaufgabe
-   * @returns {Int}
+   * @returns {Long}
    * @override
    */
-  override fun calcPartOne(): Int {
-    return 0
+  override fun calcPartOne(): Long {
+    return 0L
   }
 
   /**
    * Berechnung der zweiten Teilaufgabe
-   * @returns {Int}
+   * @returns {Long}
    * @override
    */
-  override fun calcPartTwo(): Int {
-    return 0
+  override fun calcPartTwo(): Long {
+    return 0L
   }
 }
